Code style change: Only use ternary for assignment These methods were using the ternary operator to determine which value to return. For consistency across the codebase, they have been changed to be more explicit, since the ternary operator is intended for assignment of variables.

+++ src/Builder/Delete.php
@@ -56,10 +56,14 @@ class Delete extends Builder
     /**
      * Compiles the delete from clause
      * 
-     * @return string
+     * @return null|string
      */
-    protected function compileFrom(): string
+    protected function compileFrom(): ?string
     {
+        if (! $this->table) {
+            return null;
+        }
+
         return "delete from $this->table";
     }
 }
